refactor(places-sorting): переносит логику компонента в хук
