+++ src/components/places-sorting/places-sorting.tsx
@@ -1,7 +1,6 @@
-import { KeyboardEvent, useEffect, useState } from 'react';
 import { SORTING_OPTIONS } from '../../const';
 import { SortingOption } from '../../types/sorting-option';
-import { UseBoolean } from '../../hooks/use-boolean';
+import { useSort } from '../../hooks/use-sort';
 
 
 type PlacesSortingProps = {
@@ -9,34 +8,7 @@ type PlacesSortingProps = {
 };
 
 function PlacesSorting({getSortingOption}: PlacesSortingProps): JSX.Element {
-  const { isOn, off, toggle } = UseBoolean(false);
-  const [currentOption, setCurrentOption] = useState(SORTING_OPTIONS[0] as SortingOption);
-
-  const optionChangeHandler = (option: SortingOption) => {
-    setCurrentOption(option);
-    getSortingOption(option);
-  };
-
-  const renderOptionListHandler = () => {
-    toggle();
-  };
-
-  useEffect(() => {
-    if (isOn) {
-      const escKeyDownHandler = (evt: KeyboardEvent) => {
-        if (evt.key === 'Escape') {
-          evt.preventDefault();
-          off();
-        }
-      };
-
-      document.addEventListener('keydown', () => escKeyDownHandler);
-
-      return () => {
-        document.removeEventListener('keydown', () => escKeyDownHandler);
-      };
-    }
-  }, [isOn, off]);
+  const {isOn, currentOption, optionChangeHandler, renderOptionListHandler} = useSort(getSortingOption);
 
   return(
     <form
